Remove Girder support
