+++ dandi/utils.py
@@ -625,15 +625,7 @@ def get_instance(dandi_instance_id):
         )  # note: somehow was ending up with {"girder": None}
         for name, rec in server_info.get("services", {}).items()
     }
-    if services.get("girder"):
-        return dandi_instance(
-            metadata_version=0,
-            girder=services.get("girder"),
-            gui=services.get("webui"),
-            redirector=redirector_url,
-            api=None,
-        )
-    elif services.get("api"):
+    if services.get("api"):
         return dandi_instance(
             metadata_version=1,
             girder=None,
